+++ plugins/postcss-custom-properties/CHANGELOG.md
@@ -1,6 +1,11 @@
-# 0.2.0 - 2014-08-212
+# 0.3.0 - 2014-08-25
 
-* Add map option
+* fallback now are always added by default ([see why](http://www.w3.org/TR/css-variables/#invalid-variables))
+* `map` option renamed to `variables`
+
+# 0.2.0 - 2014-08-22
+
+* Add `map` option
 * GNU style error message
 
 # 0.1.0 - 2014-08-01